+++ lib/control/flight_estimation.h
@@ -1,3 +1,5 @@
+#ifndef FLIGHT_ESTIMATION_H
+#define FLIGHT_ESTIMATION_H
 
 #include <stdio.h>
 
@@ -9,58 +11,7 @@
 
 // Status fp_update(SensorData* data, GpsData* gpsData,
 //                 FlightPhase* s_flight_phase, StateEst* currentState) {
-Status fp_update(SensorData* data, FlightPhase* s_flight_phase,
-                 StateEst* currentState) {
-    float ax = currentState->accBody.x;
-    float az = currentState->accBody.z;
-    Accel* aPtr = &data->accel;
-    void* vecPtr = (void*)aPtr;
-    (*currentState).accBody = vscale(createVectorFromStruct(vecPtr), 9.81);
-    switch (*s_flight_phase) {
-        case FP_INIT:
-            // s_flight_phase = fp_init();
-            // x is up!!!
-            if (data != NULL) {
-                *s_flight_phase = FP_READY;
-            }
-            break;
-        case FP_READY:
-            break;
-            if (ax > 15) {
-                *s_flight_phase = FP_BOOST;
-            }
-        case FP_BOOST:
-            if ((ax <= 300) && (ax >= 200)) {
-                *s_flight_phase = FP_FAST;
-            }
-            break;
-        case FP_FAST:
-            if (ax <= 200) {
-                *s_flight_phase = FP_COAST;
-            }
-            break;
-        case FP_COAST:
-            if (az >= MAINHEIGHT) {
-                if (az <= 40) {
-                    *s_flight_phase = FP_DROGUE;
-                }
-            }
-            break;
-        case FP_DROGUE:
-            if (az <= MAINHEIGHT) {
-                *s_flight_phase = FP_MAIN;
-            }
-            break;
-        case FP_MAIN:
-            if (az <= 50) {
-                *s_flight_phase = FP_LANDED;
-            }
-            break;
-        case FP_LANDED:
-            *s_flight_phase = fp_landed();
-            break;
-        default:
-            *s_flight_phase = fp_init();
-            break;
-    }
-}
+SensorData sensorFrame2SensorData(SensorFrame frame);
+void fp_update(SensorData data, FlightPhase* s_flight_phase,
+               StateEst* currentState);
+#endif  // FLIGHT_ESTIMATION_H
