+++ platformio.ini
@@ -74,6 +74,8 @@ build_flags = ${env.build_flags}
     -Isrc/pal_9k4/FreeRTOS/Source/CMSIS_RTOS_V2
     -Isrc/pal_9k4/FreeRTOS/Source/include
     -Isrc/pal_9k4/FreeRTOS/Source/portable/GCC/ARM_CM4F
+    -I.pio/build/pal_9k4/nanopb/generated-src
+    -I.pio/libdeps/pal_9k4/Nanopb
     -Isrc/pal_9k4/peripherals
     -DHSE_VALUE=48000000
     -DUSE_FULL_LL_DRIVER
